Anime Query functionality added
MALClient.cs
-> AnimeQuery method was added.
-> GetAnimeFromEntry method created: This method creates a MALAnime
object using the data given as the parameter 'element'
-> StringToQueryParameter method added: this converts a normal string to
one that could be used in the request .

MALUser.cs
-> A constructor that doesn't require user credentials added. This will
also protect any developers account from being hacked, or stolen. Use
this project as a reference in your test project, that is on a directory
seperate from the git directory.
-> Authenticate user has been made public
-> setters addded to username and passworrd properties